backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.9
gunicorn==20.1.0
